# Custom sub menu for pages
Let's say we have a module with a bunch of settings panels (general, seo, social media etc). Our client wants speed. That is why we propose this solution.

It could be admin sub sub menu, but in case we want to use custom link there should be modification. This way we have more freedom.